requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pip==19.3.1
python-crontab==2.4.0
python-dateutil==2.8.1
setuptools==42.0.2
six==1.13.0
tokenizer==2.0.3
Werkzeug==0.16.0
